+++ PlatformShared/Database/Models/SettingCrm.cs
@@ -95,7 +95,7 @@
         /// Order daily report start hour, 0-23
         /// 订单日报开始小时
         /// </summary>
-        public short? OrderDailyReportHour { get; set; }
+        public byte? OrderDailyReportHour { get; set; }
 
         /// <summary>
         /// Organization
